move manifest back to output folder
